Move files around for rpc module
So it's easy to see which methods exist
