+++ caldir-core/src/rpc/methods/create_event.rs
@@ -1,18 +1,18 @@
-use super::{Method, Rpc};
+use crate::rpc::{Method, Rpc};
 use crate::{Event, RemoteConfigParams};
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
-pub struct DeleteEvent {
+pub struct CreateEvent {
     #[serde(flatten)]
     pub remote: RemoteConfigParams,
     #[serde(with = "crate::rpc::ics")]
     pub event: Event,
 }
 
-impl Rpc for DeleteEvent {
-    const METHOD: Method = Method::DeleteEvent;
-    type Response = ();
+impl Rpc for CreateEvent {
+    const METHOD: Method = Method::CreateEvent;
+    type Response = Event;
 }
 
 #[cfg(test)]
@@ -21,7 +21,7 @@ mod tests {
     use crate::{RemoteConfigParams, event::EventTime};
 
     #[test]
-    fn delete_event_serializes_json() {
+    fn create_event_serializes_json() {
         let mut params = RemoteConfigParams::new();
         params.insert(
             "hooli_account".to_string(),
@@ -31,19 +31,16 @@ mod tests {
         let event = Event::new(
             "Test",
             EventTime::Date(chrono::NaiveDate::from_ymd_opt(2026, 1, 1).unwrap()),
-        )
-        .add_x_property("X-HOOLI-EVENT-ID", "abc123@hooli.com");
+        );
 
-        let uid = event.uid.as_str().to_string();
-
-        let cmd = DeleteEvent {
+        let cmd = CreateEvent {
             remote: params,
-            event,
+            event: event.clone(),
         };
 
         let json = cmd.to_json().unwrap();
 
-        assert_eq!(json["command"], "delete_event");
+        assert_eq!(json["command"], "create_event");
         assert_eq!(json["params"]["hooli_account"], "user@hmail.com");
 
         let ics = json["params"]["event"]
@@ -51,7 +48,6 @@ mod tests {
             .expect("event should be a string");
 
         assert!(ics.starts_with("BEGIN:VCALENDAR"));
-        assert!(ics.contains(&format!("UID:{}", uid)));
-        assert!(ics.contains("X-HOOLI-EVENT-ID:abc123@hooli.com"));
+        assert!(ics.contains(&format!("UID:{}", event.uid.as_str())));
     }
 }
